Processing: Extract highwayAreas predicate without table require
- Tests can check highway=* + area=yes without loading osm2pgsql table defs
Ping https://github.com/FixMyBerlin/tilda-geo/pull/379
