+++ processing/topics/roads_bikelanes/roads_bikelanes.lua
@@ -11,11 +11,12 @@ local roads_bikelanes_bikelanes = require('topics.roads_bikelanes.roads_bikelane
 local roads_bikelanes_roads = require('topics.roads_bikelanes.roads_bikelanes_roads')
 local EXIT = require('topics.roads_bikelanes.helper.exit_processing')
 local result_tags = require('topics.roads_bikelanes.helper.result_tags')
+local is_highway_area = require('topics.roads_bikelanes.highway_areas.is_highway_area')
 local roads_bikelanes_highway_areas = require('topics.roads_bikelanes.roads_bikelanes_highway_areas')
 
 ---@param object table
 function osm2pgsql.process_way(object)
-  if object.is_closed and object.tags.highway and object.tags.area == 'yes' then
+  if is_highway_area(object) then
     roads_bikelanes_highway_areas.process_highway_area(object)
     return
   end
@@ -59,7 +60,7 @@ function osm2pgsql.process_way(object)
 end
 
 function osm2pgsql.process_relation(object)
-  if object.tags.highway and object.tags.area == 'yes' then
+  if is_highway_area(object) then
     roads_bikelanes_highway_areas.process_highway_area(object)
   end
 end
